Fix first round after a restart being displayed as number 0
